Split tests to own files
